logo_dl/pyproject.toml
[tool.poetry]
name = "logo-dl"
version = "0.1.0"
description = ""
authors = ["tigattack <10629864+tigattack@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiofiles = "^24.1.0"
aiohttp = "^3.10.5"
beautifulsoup4 = "^4.12.3"
pydantic = "^2.9.1"
python-slugify = "^8.0.4"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.2"
ruff = "^0.6.4"
types-aiofiles = "^24.1.0.20240626"
types-beautifulsoup4 = "^4.12.0.20240907"
types-requests = "^2.32.0.20240907"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
